PlayerToggleSneakListener.java implemented
